Add files via upload
The application takes the user name to enter into the quiz. Number of questions is to be answered. On clicking on the answers, correct and wrong answers are notified, Once the quiz is answered, the user score is shown. On clicking the view leaderboard button, leaderboard is displayed with the rankings.